+++ src/components/features/cart/cartItem/CartItem.tsx
@@ -1,9 +1,10 @@
-import { refetchData } from '@/shared/data/dataStore';
-import { useJaeOMutation } from '@/shared/data/useJaeOMutation';
-import { showErrorToast } from '@/shared/toast/toastStore';
-import { isValidImageUrl } from '../../../../shared/utils/isValidImageUrl';
-import SelectBox from '../../../common/selectBox/SelectBox';
-import Separator from '../../../common/separator/Separator';
+import { SelectBox, Separator } from '@/components/common';
+import {
+  isValidImageUrl,
+  refetchData,
+  showErrorToast,
+  useJaeOMutation,
+} from '@/shared';
 import { deleteCartItem } from '../api/deleteCartItem';
 import { updateCartItem } from '../api/updateCartItem';
 import CartQuantityControlButton from '../cartQuantityControlButton/CartQuantityControlButton';
