+++ src/components/features/cart/cartTitle/CartTitle.tsx
@@ -1,4 +1,4 @@
-import PageTitle from '@/components/common/pageTitle/PageTitle';
+import { PageTitle } from '@/components/common';
 
 interface CartTitleProps {
   quantity?: number;
