+++ src/shared/hooks/useToggle.ts
@@ -1,9 +1,7 @@
 import { useState } from 'react';
 
-function useToggle(initial = false) {
+export function useToggle(initial = false) {
   const [state, setState] = useState(initial);
   const toggle = () => setState((prev) => !prev);
   return [state, toggle, setState] as const;
 }
-
-export default useToggle;
